[Server] ADD 북마크, 리뷰 API 등.

+++ server/modules/image.js
@@ -1,6 +1,8 @@
 const uuid = require('uuid').v4;
 const AWS = require('aws-sdk');
-const debug = require('debug')('app');
+const debug = require('debug')('app:image');
+const fs = require('fs').promises;
+const { Client } = require('node-scp');
 
 const v4 = () => uuid().replace(/-/g, '').slice(0, 15);
 
@@ -12,7 +14,7 @@ const {
 } = process.env;
 
 // Configure AWS to use promise
-AWS.config.setPromisesDependency(require('bluebird'));
+// AWS.config.setPromisesDependency(require('bluebird'));
 
 AWS.config.update({
   accessKeyId: ACCESS_KEY_ID,
@@ -30,6 +32,32 @@ module.exports = {
     const base64Data = Buffer.from(matches[2], 'base64');
     const type = matches[1];
     const uid = v4();
+
+    // 개발모드일 경우 테스트서버에 저장
+    if (process.env.NODE_ENV !== 'production') {
+      const filename = `${uid}.${type.split('/')[1]}`;
+      await fs.mkdir(`/tmp/${path}`, { recursive: true });
+      await fs.writeFile(`/tmp/${path}/${filename}`, base64Data);
+      // SCP로 전송
+      try {
+        const client = await Client({
+          host: process.env.TEST_SSH_HOST,
+          port: parseInt(process.env.TEST_SSH_PORT, 10),
+          username: process.env.TEST_SSH_USERNAME,
+          password: process.env.TEST_SSH_PASSWORD,
+        });
+        await client.uploadDir(`/tmp/${path}`, `/web/gunsigi_data/images/${path}`);
+        client.close();
+      } catch (e) {
+        debug(e);
+        throw Error('Error in save image');
+      } finally {
+        await fs.rm(`/tmp/${path}`, { recursive: true });
+      }
+      return `${path}/${filename}`;
+    }
+
+    // S3저장
     const params = {
       Bucket: BUCKET,
       Key: `${path}/${uid}`,
@@ -48,10 +76,12 @@ module.exports = {
     return key;
   },
   delete: async (Key) => {
+    if (process.env.NODE_ENV !== 'production') return;
     // CDN서버에 있는 이미지를 삭제한다.
     await s3.deleteObject({ Bucket: BUCKET, Key }).promise();
   },
   deleteFolder: async function deleteObjects(dir) {
+    if (process.env.NODE_ENV !== 'production') return;
     // Objects 조회
     const listParams = { Bucket: BUCKET, Prefix: dir };
     const listObjects = await s3.listObjectsV2(listParams).promise();
